Deprecation: Quoting % in YAML Most of the other deprecations are pretty easy, like the next one:
Not quoting the scalar %cache_type% starting with the % indicator character is deprecated since Symfony 3.1.

This, and the next deprecation are in config.yml - and it even tells us the exact lines!

Open up app/config/config.yml and find line 71. Yep! Put quotes around %cache_type%. To more closely follow the official YAML spec, if a value starts with %, it needs to have quotes around it. Do the same around the directory value.