Where else are we using {dollar_sign}container->get()? Let's find out! In your terminal, run:
git grep '\->get('
Ah, just two more! We may not need to change all of these... some services are still public. But let's clean it all up!

Start in SecurityController. Ah, here it is. So: what type-hint should we use to replace this? Well, you could just guess! Honestly, that works a lot. Or try the brand new console command:

./bin/console debug:autowiring
Sweet! This is a full list of all type-hints you can use for autowiring. Search for 'authentication' and... there it is! This type-hint is an alias to the service we want.
That means, back in SecurityController, delete this line and add a new AuthenticationUtils argument. Done.
The last spot is in UserController: we're using security.authentication.guard_handler. This time, let's guess the type-hint! Add a new argument: Guard... GuardAuthenticationHandler. That's probably it! And if we're wrong, Symfony will tell us. Use that value below.