Our project is now on Symfony 4.0, and it still works! Well, it almost works: we would just need to remove a few references to SensioDistributionBundle and SensioGeneratorBundle.
The point is this: if you want, you can upgrade to Symfony 4, but not migrate to the new Flex project structure. That's fine.

But... since Flex is awesome... let's do it!

Flex: Composer Plugin & New BFF
Flex is a Composer plugin, and, it's pretty simple: when you install a package, it checks to see if there is a recipe for that package. A recipe can add configuration files, auto-enable the bundle, add paths to your .gitignore file and more. But, for Flex to work, you need to use the Flex directory structure.

Upgrade to Flex: The Plan
So here's the plan: we're going to bootstrap a new Flex application right inside our existing project. Then, little-by-little, we'll move our code and configuration into it. It's going to be pretty freakin' cool.

Upgrade Composer. For Real
Before we start, make sure that your Composer is at the latest version:

composer self-update
Seriously, do this. Composer recently released a bug fix that helps Flex.