We need to make our composer.json file look like the one from symfony…
…/skeleton. Actually, navigate to https://github.com/symfony/skeleton, go to 'Releases', find the latest release, and then click to browse the files. Now we can see the stable composer.json contents.

So... yea, this one file is all you need to start a new project. That's crazy! Flex builds the project structure around it.

Bye Bye symfony/symfony
Anyways, the most important change is that, with Flex, you stop requiring symfony/symfony. Yep, you require only the specific packages that you need. Copy all of the require lines, find our composer.json file, and paste over the php and symfony/symfony lines. Oh, and remove symfony/flex from the bottom: it's up here now. The symfony/framework-bundle package is the most important: this is the core of Symfony: it's really the only required package for a Symfony app.

Go back and also copy the dotenv package from require-dev and put it in our composer.json file. This package is responsible for reading the new .env file.

Synchronizing the rest of Composer.json
Go back and also copy the config line and paste that here too. Skip the autoload sections for now, but copy the rest of the file. Replace the existing scripts and extras sections with this new, shiny stuff.