Autoloading src/ & src/AppBundle
Let's talk about autoload. In Symfony 3, everything lived in src/AppBundle and had an AppBundle namespace. But in Symfony 4, as you can see, everything should live directly in src/. And even though we don't have any examples yet, the namespace will start with App\, even though there's no App/ directory.

Eventually, we are going to move all of our files into src/ and refactor all of the namespaces. With PhpStorm, that won't be as scary as you think. But, it is a big change, and you may not be able to do that all at once in a real project.

So, I'm going to show you a more gentle', gradual way to upgrade to Flex. Yep, for now, we're going to leave our files in AppBundle and make them work. But new files will live directly in src/.

Right now, the autoload key in composer.json says to look for all namespaces in src/. Make this more specific: the AppBundle\ namespace prefix should live in src/AppBundle. Do the same in autoload-dev: Tests\AppBundle\ will live in tests/AppBundle.
Why are we doing this? Because now we can go copy the autoload entry from the official composer.json file and add it below our AppBundle\ line. Copy the new autoload-dev line also.
Thanks to this, Composer can autoload our old classes and any new classes!

Scaffold the Full Structure