Thanks to the Flex recipe for symfony/framework-bundle, we now have a fully-functional Symfony Flex app living right inside our directory! public. is the new document root, config/ has all of the configuration, and our PHP code lives in src/, including the new Kernel class.
Yep, we have our old app with all our stuff, and a new, Flex, app, which is basically empty and waiting for us to move our code into it.

Re-Order .env
Open up .env.dist. Woh! This has more stuff now! That's thanks to the recipes from DoctrineBundle, SwiftmailerBundle and FrameworkBundle. Copy the FrameworkBundle section and move that to the top. Do the same thing to .env.
We don't need to do this, but APP_ENV is so important, I want to see it first. If you start a new Flex app, it's on top.