Re-Ordering the Libs Next, this will sound weird, but run:
composer require symfony/flex
We already have this library. I know. So... why are we doing this? It's a little trick: one of the new keys in our composer.json is sort-packages, which is set to true. Thanks to this, whenever you run composer require, it orders the packages alphabetically. By requiring a package we already have, Composer just re-ordered my packages.