Fixing the console But... we still have this giant error: attempted to load SecurityBundle from AppKernel. Bummer! This happens because bin/console is still trying to boot the old app.
When you start a new Flex project, the symfony/console recipe creates the bin/console file. But, since our project already had this file, the recipe couldn't do its job.

No worries! Let's go find the new file! Go to github.com/symfony/recipes. Welcome to the official recipes repository!

Navigate to symfony, console, then bin. There it is! Copy its contents. Then, completely replace our version.

This will boot the new application! So... does it work? Run:

./bin/console
No! But that's a new error: we are closer! This says that the autoloader expects App\AppBundle\AppBundle to be defined in AppBundle.php, but it wasn't found. That's strange... that is not the correct namespace for that class! If you look closer, it says the error is coming from a new config/services.yaml file.