Requiring translator The next key is translator. Are you ready? Because this is where things get fun! You might think that all we need to do is copy this line into framework.yaml. But no!
Many of the keys under framework represent components. In Symfony 3, by adding the translator key, you activated that component.

But with Flex, the Translator component isn't even installed yet. Yep, if you want a translator, you need to install it. In your terminal, run:

composer require translator
If that package name looks funny... it should! There is no package called translator! But look! It added a new symfony/translator key to composer.json.

This is another superpower of Flex. Go to https://symfony.sh/. This is a list of all of the packages that have a recipe. Search for translation to find symfony/translation. See those Aliases? Yep, we can reference translation, translations or translator and Flex will, um, translate that into symfony/translator automatically.

The translator Recipe
Back to the terminal! Before I started recording, I committed all of our changes so far. That was no accident: Flex just installed a recipe and I want to see exactly what it did! Run:

git status
Cool! It created a new translation.yaml file and a translation/ directory.