So, do we need to move these form and validation lines of config into framework.yaml? Actually, no!
Go back to your terminal and run:

./bin/console debug:config framework
This prints out the current framework configuration. Search for form. Nice! It's already enabled, even without any config! This is really common with Flex: as soon as a component is installed, FrameworkBundle automatically enables it. No configuration is needed unless you want to change something. Delete the form line.

Search for 'validation' next: it's even more interesting! It's also enabled, and enable_annotations is set to true. Great! Delete the validation line! What's really interesting is that enable_annotations is set to true because it detected that we have the Doctrine annotations package installed. This is the flow with Flex: install a package and you're done.

Ok! It might not look like it, but we're almost done with the framework stuff. Let's finish it next!