Previously onAuthenticationSuccess method was handled by the base class for you. But now, it's your responsibility. No worries: it's pretty simple. To help, at the top, use a trait called TargetPathTrait. Back down in onAuthenticationSuccess, this allows us to say if {dollar_sign}targetPath = {dollar_sign}this->getTargetPath() with {dollar_sign}request->getSession() and main.
Let's break this down. First, the main string is just the name of our firewall. In both the old and new security config, that's its key.

Second, what does getTargetPath() do? Well, suppose the user originally tried to go to /admin, and then they were redirected to the login page. After they login, we should probably send them back to /admin, right? The getTargetPath() method returns the URL that the user originally tried to access, if any.

So if there is a target path, return new RedirectResponse({dollar_sign}targetPath). Else, return new RedirectResponse and generate a URL to the homepage.

PhpStorm thinks this isn't a real route, but it is!

Problem solved! Is that enough to make our app happy? Find out!

./bin/console
It is!