We are in the Flex home stretch! These last config/ files are the easiest. Start with config_dev.yml.
Dev Environment Parameters
Ok, we have a cache_type parameter. This is meant to override the value that lives in services.yaml whenever we're in the dev environment.

How can we have dev-specific parameters or services in Flex? By creating a new services_dev.yaml file. Copy the parameter, remove it and paste it here.
Symfony will automatically load this file in the dev environment only.