Moving & Delete Files Let's reset our files and look at each directory one by one. We need bin/, config/ and public/ is the new document root. src/ holds our code, and templates/, tests/ and translations/ are all valid Flex directories. Oh, and tutorial/? Ignore that: I added that for this course - it has a file we'll use later.
But expand var/. Delete everything except for cache and log: the default logs directory was renamed in Flex. And that bootstrap file is a relic of the past!

And finally.... web/! This directory should not exist. Select the files we need: css/, images/, js/ and vendor/: move these into public/. Let's also move robots.txt.

And that's it! The favicon is from Symfony and we don't need the app files anymore. What about .htaccess? You only need that if you use Apache. And if you do, Flex can add this file for you! Just run composer require symfony/apache-pack. The recipe will add this inside public.

Anyways, delete web/! This is it! Our app is fully in Flex! And we didn't even need to move all our files from src/AppBundle... though we will do that soon. And as far as bin/console is concerned, the app works!

But to really prove it's alive, let's try this in a browser and handle a few last details. That's next!