If you want to stop now, you can! Your old code lives in src/AppBundl…
…e, but it works! Over time, you can slowly migrate it directly into src/.

Or! We can keep going: take this final challenge head-on and move all our files at once! If you're not using PhpStorm... this will be a nightmare. Yep, this is one of those rare times when you really need to use it.

Moving your Files
Open AppBundle.php. Then, right click on the AppBundle namespace and go to Refactor -> Move. The new namespace will be App. And below... yea! The target destination should be src/.

This says: change all AppBundle namespaces to App and move things into the src/ directory. Try it! On the big summary, click OK!

In addition to changing the namespace at the top of each file, PhpStorm is also searching for references to the namespaces and changing those too. Will it be perfect? Of course not! But that last pieces are pretty easy.

Woh! Yes! Everything is directly in src/. AppBundle is now empty, except for a fixtures.yml file. We're going to replace that file soon anyways.

Delete AppBundle! That felt amazing!

Refactoring tests/
Let's do the same thing for the tests/ directory... even though we only have one file. Open DefaultControllerTest.php and Refactor -> Move its namespace. In Flex, the namespace should start with App\Tests. Then, press F2 to change the directory to tests/Controller.
Ok, Refactor! Nice! Now delete that AppBundle.