Start in security.yaml and do the same find and replace. You could do this for your entire project, but I'll play it safe. Now, completely delete the AppBundle.php file: we're already not using that. Next is GenusAdminController. Open that class. But instead of replacing everything, which would work, search for AppBundle. Ah! It's a getRepository() call!
Our project has a lot of these... and... well... if you're lazy, there's a secret way to fix it! Just change the alias in doctrine.yaml from App to AppBundle. Cool... but let's do it the right way! Use Genus::class.

We have a few more in GenusController. Use SubFamily::class, User::class, Genus::class, GenusNote::class and GenusScientist::class.

Ok, back to the list! Ah, a few entities still have AppBundle. Start with Genus. The repositoryClass, of course! Change AppBundle to App. There's another reference down below on a relationship. Since all the entities live in the same directory, this can be shortened to just SubFamily.

Make the same change in GenusNote, SubFamily and User.