Almost done! Next is GenusFormType: open that and change the data_class to Genus::class. Then, finally, LoginFormAuthenticator. Update AppBundle:User to User::class.
Phew! Search for AppBundle again:

git grep AppBundle
They're gone! So... ahh... let's try it! Refresh! Woh! An 'Incomplete Class' error? Fix it by manually going to /logout. What was that? Well, because we changed the User class, the User object in the session couldn't be deserialized. On production, your users shouldn't get an error, but they will likely be logged out when you first deploy.

Go back to /admin/genus, then login with weaverryan+1@gmail.com, password iliketurtles. Guys, we're done! We have a Symfony 4 app, built on the Flex directory structure, and with no references to AppBundle! And it was all done in a safe, gradual way.

To celebrate, I've added one last video with a few reasons to be thrilled that you've made it this far.