Now that we're on Symfony 4 with Flex, I have three cool things to show you.
Repositories as a Service
Start by opening GenusController: find listAction. Ah yes: this is a very classic setup: get the entity manager, get the repository, then call a method on it.

One of the annoying things is that - unless you add a bunch of extra config - repositories are not services and can not be autowired. Boo!

Well... that's not true anymore! Want your repository to be a service? Just make two changes. First, extend a new base class: {dollar_sign}erviceEntityRepository.

And second, override the __construct() function. But remove the {dollar_sign}entityClass argument. In the parent call, use Genus::class.

That might look weird at first... but with those two small changes, your repository is already being auto-registered as a service! Yep, back in listAction, add a new argument: GenusRepository {dollar_sign}genusRepository. Use that below instead of fetching the EntityManager.
And that's it! Go to that page in your browser: /genus. Beautiful! Make that same change to your other repository classes when you want to.