Ready for one last cool thing? Run:
composer require maker

This installs the MakerBundle: Symfony's new code generator. Code generation is of course optional. But with this bundle, you'll be able to develop new features faster than ever. Need a console command, an event subscriber or a Twig extension? Yep, there's a command for that.

What's everything it can do? Run:

./bin/console list make

Right now, it has about 10 commands - but there are a lot more planned: this bundle is only about 1 month old!